requirements/requirements.txt
Django==4.2
pyaml==21.10.1
